Port to using the newer graph, which offers iterators instead of the older `each` method, but is otherwise identical.
