+++ src/librustc_data_structures/lib.rs
@@ -33,3 +33,5 @@
 extern crate serialize as rustc_serialize; // used by deriving
 
 pub mod snapshot_vec;
+pub mod graph;
+pub mod bitvec;
